test for communication

+++ siemens_plc/scripts/siemens_plc_interface.py
@@ -37,7 +37,8 @@ def laser_status_callback(msg, args):
     
     # get the laser_status msgs and update the laser_config array
     rospy.loginfo("Laser status from modbus server: %s", str(msg.data))
-    args[1].data = [msg.data[0], np.int32(ready_flag), np.int32(powder_feed_start), np.int32(blow_gas_start), np.int32(emit_laser_start), np.int32(laser_power), np.int32(powder_feed_rate)]
+    args[1].data = [msg.data[0], ready_flag, powder_feed_start, blow_gas_start, emit_laser_start, laser_power, powder_feed_rate]
+    # args[1].data = [msg.data[0], np.int32(ready_flag), np.int32(powder_feed_start), np.int32(blow_gas_start), np.int32(emit_laser_start), np.int32(laser_power), np.int32(powder_feed_rate)]
     
     # publish the laser config parameters
     args[0].publish(args[1])
